+++ src/main/java/com/fullstack/stock/Controller/ProdutoController.java
@@ -8,6 +8,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigInteger;
 import java.net.URI;
 import java.util.List;
 
@@ -48,7 +49,7 @@ public class ProdutoController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(@PathVariable BigInteger id) {
         produtoService.delete(id);
         return ResponseEntity.noContent().build();
     }
